requirements.txt
flask
numpy
pillow
pydicom
pycryptodome
python-dotenv
matplotlib
seaborn
scikit-image
